Add files via upload
This was the first program written to analyze the clean data from Stata. The initial analyses did not age-standardize the prevalence estimates. The research team made a decision in September 2021 to present age-standardized weighted prevalence estimates. This file was archived on September 9, 2021 and a new R file created for the new analyses.